simplify execute quote file, clean up
