--- packages/sdk/test/unit/actions/client.test.ts
@@ -1,11 +1,0 @@
-import { describe, expect, test } from "vitest";
-import { testClient } from "../../common/client";
-import { DefaultLogger } from "../../../src";
-
-describe("Initialize client", () => {
-  test("Client configured correctly", () => {
-    expect(testClient).to.not.be.undefined;
-    expect(testClient.logger).toBeInstanceOf(DefaultLogger);
-    expect(testClient.isTenderlyEnabled).toBe(false);
-  });
-});
